+++ lib/vcloud/launcher/vapp_orchestrator.rb
@@ -12,7 +12,8 @@ module Vcloud
         end
 
         template_name = vapp_config[:vapp_template_name] || vapp_config[:catalog_item]
-        template = Vcloud::Core::VappTemplate.get(template_name, vapp_config[:catalog])
+        catalog_name = vapp_config[:catalog_name] || vapp_config[:catalog]
+        template = Vcloud::Core::VappTemplate.get(template_name, catalog_name)
         template_id = template.id
 
         network_names = extract_vm_networks(vapp_config)
@@ -30,7 +31,8 @@ module Vcloud
           internals: {
             name:               { type: 'string', required: true, allowed_empty: false },
             vdc_name:           { type: 'string', required: true, allowed_empty: false },
-            catalog:            { type: 'string', required: true, allowed_empty: false },
+            catalog:            { type: 'string', deprecated_by: 'catalog_name', allowed_empty: false },
+            catalog_name:       { type: 'string', required: true, allowed_empty: false },
             catalog_item:       { type: 'string', deprecated_by: 'vapp_template_name', allowed_empty: false },
             vapp_template_name: { type: 'string', required: true, allowed_empty: false },
             vm: Vcloud::Launcher::VmOrchestrator.customize_schema,