+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Maint:
 
   - Deprecate 'catalog_item' for 'vapp_template_name' in config.
+  - Deprecate 'catalog' for 'catalog_name' in config.
 
 ## 0.0.5 (2014-05-14)